==> notebooks/1-preprocessing.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

!pip install openai
!pip install huggingface_hub
!pip install transformers
!pip install datasets

import os
import pandas as pd
import numpy as np
from datasets import load_dataset
import re
import nltk
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer
nltk.download('punkt')
nltk.download('stopwords')
nltk.download('wordnet')

hf_token = os.getenv("llm_project_token")

from datasets import load_dataset
ds = load_dataset("stanfordnlp/imdb", token=hf_token)

ds_train = pd.DataFrame(ds['train'])
ds_test = pd.DataFrame(ds['test'])
ds_unsupervised = pd.DataFrame(ds['unsupervised'])

### **Exploring the Dataset**

ds

ds['train'][0]

ds['train'].features

ds_train = pd.DataFrame(ds['train'])
ds_test = pd.DataFrame(ds['test'])

from datasets import Dataset, DatasetDict
train = Dataset.from_pandas(ds_train)
test = Dataset.from_pandas(ds_test)
new_ds = DatasetDict(
    {
        'train': train,
        'test': test
    }
)
new_ds

#Checking for nulls
ds_train.isnull().sum()

ds_test.isnull().sum()

### **Text Cleaning**

stop_w = set(stopwords.words('english'))
lemmatizer = WordNetLemmatizer()

def clean_text(text):
    text = re.sub(r'<.*?>', '', text)
    text = re.sub(r'[^a-zA-Z\s]', '', text)
    words_t = word_tokenize(text)
    words_t = [lemmatizer.lemmatize(word) for word in words_t if word.lower() not in stop_w]
    return ' '.join(words_t)

ds_train['clean_text'] = ds_train['text'].apply(clean_text)
ds_test['clean_text'] = ds_test['text'].apply(clean_text)
ds_unsupervised['clean_text'] = ds_unsupervised['text'].apply(clean_text)

### **Saving Dataset**

ds_train.to_csv('/content/drive/MyDrive/Colab Notebooks/ds_train.csv', index=False)
ds_test.to_csv('/content/drive/MyDrive/Colab Notebooks/ds_test.csv', index=False)
