==> notebooks/2-representation.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from sklearn.svm import SVC
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, confusion_matrix, classification_report

ds_train = pd.read_csv ('/content/drive/MyDrive/Colab Notebooks/ds_train.csv')
ds_test = pd.read_csv ('/content/drive/MyDrive/Colab Notebooks/ds_test.csv')

y_train = ds_train['label']
y_test = ds_test['label']

### **TD-IDF and BoW Representations**

tfidf_v = TfidfVectorizer()
bow_v = CountVectorizer()


X_train_t = tfidf_v.fit_transform(ds_train['clean_text'])
X_test_t = tfidf_v.transform(ds_test['clean_text'])

X_train_b = bow_v.fit_transform(ds_train['clean_text'])
X_test_b = bow_v.transform(ds_test['clean_text'])

def te_model(model, X_train, y_train, X_test, y_test):
    model.fit(X_train, y_train)
    y_pred = model.predict(X_test)

    accuracy_s = accuracy_score(y_test, y_pred)
    precision_s = precision_score(y_test, y_pred)
    recall_s = recall_score(y_test, y_pred)
    f1_s = f1_score(y_test, y_pred)
    confusion_m = confusion_matrix(y_test, y_pred)
    c_report = classification_report(y_test, y_pred)

    print(f'Accuracy: {accuracy_s}')
    print(f'Precision: {precision_s}')
    print(f'Recall: {recall_s}')
    print(f'F1 Score: {f1_s}')
    print('Confusion Matrix:')
    print(confusion_m)
    print('Classification Report:')
    print(c_report)

    return accuracy_s, precision_s, recall_s, f1_s, confusion_m, c_report

### **TF-IDF Model Training**

print("TF-IDF Models")

# Logistic Regression
print("Logistic Regression")
logr_model = LogisticRegression()
te_model(logr_model, X_train_t, y_train, X_test_t, y_test)

# Random Forest
print("Random Forest")
ranf_model = RandomForestClassifier()
te_model(ranf_model, X_train_t, y_train, X_test_t, y_test)

# Gradient Boosting
print("Gradient Boosting")
gradb_model = GradientBoostingClassifier()
te_model(gradb_model, X_train_t, y_train, X_test_t, y_test)

# SVM
print("SVM")
svm_model = SVC(kernel='linear', probability=True)
te_model(svm_model, X_train_t, y_train, X_test_t, y_test)

### **TF-IDF Model Insights**

**Logistic Regression and SVM** appear to be the best-performing models with high accuracy, balanced precision, recall, and F1 scores.

**Random Forest** has a decent performance with a lower recall which potentially could provide less reliable results. There is a likelihood of being unable to identify positive cases properly.

**Gradient Boosting** offers the lowest performance especially in accuracy and percision. It is an unideal choice for this analysis.

### **BoW Models Training**

print("BoW Models")

# Logistic Regression
print("Logistic Regression")
logr_model = LogisticRegression(max_iter=1000)
te_model(logr_model, X_train_b, y_train, X_test_b, y_test)

# Random Forest
print("Random Forest")
ranf_model = RandomForestClassifier(n_estimators=100)
te_model(ranf_model, X_train_b, y_train, X_test_b, y_test)

# Gradient Boosting
print("Gradient Boosting")
gradb_model = GradientBoostingClassifier(n_estimators=100)
te_model(gradb_model, X_train_b, y_train, X_test_b, y_test)

# SVM
print("SVM")
svm_model = SVC(kernel='linear', probability=True)
te_model(svm_model, X_train_b, y_train, X_test_b, y_test)

### **BoW Model Insights**

**Logistic Regression** offers the best performing model with the highest accuracy and precision, recall, and F1 score.

**Random Forest and SVM** both perform well but slightly lower compared to Logistic Regression. These models would be alternative choices if needed.

**Gradient Boosting** provides the lowest performance especially in precision if focus is minimizing false positivies. However, the high recall could be beneficial if you're only looking to identify positive cases.

### **TF-IDF and BoW Comparison**

**TF-IDF Models**
These models perform better than BoW models especially SVM and Logistic Regression with great improvement in precision, recall, and F1 score.

**BoW Models**
BoW Models perform well but slightly lower performance metrics compared to TF-IDF.

Overall, TF-IDF offers more effective feature extraction method: SVM and Logistic Regression, offer better accuracy and balanced performance metrics. The choice for feature extraction will depend on the importance of precision vs recall in the application.
