==> notebooks/3-pre-trained-model.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import numpy as np
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, confusion_matrix, classification_report
from transformers import pipeline, DistilBertTokenizerFast, DistilBertTokenizer

ds_train = pd.read_csv ('/content/drive/MyDrive/Colab Notebooks/ds_train.csv')
ds_test = pd.read_csv ('/content/drive/MyDrive/Colab Notebooks/ds_test.csv')

y_train = ds_train['label']
y_test = ds_test['label']

### **Sentiment Analysis Pipeline**

tokenizer = DistilBertTokenizer.from_pretrained('distilbert-base-uncased')

### **Tokenization and Inference**

# Example sentence
sentence = 'Tokenize this sentence using DistilBERT.'

# Tokenize the sentence
tokens = tokenizer(sentence, return_tensors='pt')

# Retrieve the input_ids
input_ids = tokens['input_ids'].squeeze().tolist()

# Retrieve the attention mask
attention_mask = tokens['attention_mask'].squeeze().tolist()

# Print the results
print(f'Original sentence:\n{sentence}\n')
print(f'Input IDs:\n{input_ids}\n')
print(f'Attention mask:\n{attention_mask}')

for input_id in input_ids:
      print(f'{input_id}: {tokenizer.decode([input_id])}')

classifier = pipeline(task="sentiment-analysis", model="distilbert-base-uncased-finetuned-sst-2-english")
tokenizer = DistilBertTokenizerFast.from_pretrained("distilbert-base-uncased-finetuned-sst-2-english")

def predictions(text_list, classifier, max_length=512):
    preds = classifier(text_list, padding=True, truncation=True, max_length=max_length)
    pred_labels = [1 if pred['label'] == 'POSITIVE' else 0 for pred in preds]
    return pred_labels

train_p = predictions(ds_train['clean_text'].tolist(), classifier)
test_p = predictions(ds_test['clean_text'].tolist(), classifier)

print(train_p[:15])
print(test_p[:15])

def ev_model(y_true, y_pred):
    accuracy_s = accuracy_score(y_test, y_pred)
    precision_s = precision_score(y_test, y_pred)
    recall_s = recall_score(y_test, y_pred)
    f1_s = f1_score(y_test, y_pred)
    confusion_m = confusion_matrix(y_test, y_pred)
    c_report = classification_report(y_test, y_pred)

    print(f'Accuracy: {accuracy_s}')
    print(f'Precision: {precision_s}')
    print(f'Recall: {recall_s}')
    print(f'F1 Score: {f1_s}')
    print('Confusion Matrix:')
    print(confusion_m)
    print('Classification Report:')
    print(c_report)

    return accuracy_s, precision_s, recall_s, f1_s, confusion_m, c_report

### **Evaluation**

print("Train Dataset Evaluation:")
ev_model(y_train, train_p)

print("Test Dataset Evaluation:")
ev_model(y_test, test_p)
